Add dicom_dir_to_3d_arr and load_npz functions to utils.py
Added functions to utils.py to load a numpy array from a case directory
where each file in the directory corresponds to a dicom file of a particular
2D slice.

Also added a function which can load the npz segmentation files.

These functions are required to load the data in the format it was given
for the purposes of the coursework.

Added an auxilary function which loads the data using the above functions
and saves the resulting volume and segmentation as nifti files.

This is a useful format to load the files into popular medical image viewers
like ITKSNAP and 3D Slicer. Allows for easy visualisation of a particular case
with the segmentation loaded on top.