Add ImageSegmentationDataset class
Added Image Segmentation Dataset class to serve cases during training and
testing.

Assumes all cases are already loaded, this was a deliberate design choice
due to the structure of the data given for this coursework.
The data is 2d however given for the masks as a single volume per case, therefore
it must be loaded and batched/split outside of the dataset class.