Add Invert Mask Logic to train.py and Visualise Option in Config
Added logic to invert the binary mask so that background values are 1
and the foreground (lung) values are 0.

Update added because many slices feature only background and the dice loss
for such slices will always be 1. This results in unstable training since
model can not do backprop to change the loss for such slices.

For better readability and configurability, made the option to visualise a slice
at the end of every epoch a configurable option. This option is in addition to the
visualisations that are uploaded to weights and biases on every training job.

Also updated the epoch loop to start from 1 instead of 0.