Add test.py
Adds main testing script, goal of this script is to use a trained model to perform
inference on cases and save the dice and accuracy metrics to a csv.

All output files are saved in the output dir specified in the testing config.
Optionally, segmentations can also be saved in nifti format, for ease of viewing with
current medical image viewers like ITKSNAP and 3D Slicer.
Segmentations will be saved in a sub dir of this output dir.

Testing config used will also be automatically copied to this output directory.

No weights and biases logger used to track testing job.