Add visualise_segs.py
Quick script to visually compare the segmentations and mask of a given
case along with the dice and accuracy metrics.

Takes as input the output_dir of test.py which is expected to contain a metrics csv
file and a segmentations sub dir with nifti pred files.
Extracts slice idx and case id from the pred file name.
Loads the corresponding case from the image and mask dirs and plots the relevant
slice.

Also shows the metrics for that slice as saved in the csv file.