Refactor Train Test Splitting Logic to Prevent Data Leakage
Changed the logic by which the data is split into train and test splits
as prev method loaded slices from all cases and would potentially assign
slices from the same case to both the train and test sets.

Slices belong to a specific case are not independant.
As such, to prevent information leakage between the train and test sets
the case ids now need to be explicitly passed in the configs.

Only slices for these cases will be used to train or test a model.

This also removes the need for the test_split and eval_train_split
config options.