Add Containerisation Modifications
Updated environment yml, updated docker file, started read me.

Had to make the case ids case sensitive in yml files for it to work
in the docker container... strange how thats not needed for it to run locally.

Need to add download instructions in read me for the model weights and dataset.
Place holder lines added atm.

Need to check the testing and eval pipelines work.
Need to add instructions to activate the env and generate figures from notebook.